[add] task checking function ADD

+++ LittleAchievement/task/views.py
@@ -12,13 +12,20 @@ def subscribe(request):
         print("새로운 할일이 등록 되었습니다.")
     return redirect('index')
 
-def unsubscribe(request):
+def task_managing(request):
     active_user  = request.user
-    target_task = MyTask.objects.get(id=request.POST.get('unsubscribe_task'))
     
-    if target_task.user == active_user:
-        target_task.delete()
-        print("할일이 삭제되었습니다.")
+    if request.POST.get('unsubscribe_task'):
+        target_task = MyTask.objects.get(id=request.POST.get('unsubscribe_task'))
+        if target_task.user == active_user:
+            target_task.delete()
+            print("할일이 삭제되었습니다.")
+    if request.POST.get('checking_task'):
+        checking_task = MyTask.objects.get(id=request.POST.get('checking_task'))
+        if checking_task.user == active_user:
+            checking_task.is_checked = True
+            checking_task.save()
+            print("하나의 일을 달성 하셨군요! 수고하셨어요!!.")
 
     print("현재 등록된 Task 갯수 ; ", MyTask.objects.filter(user=active_user).count())
     return redirect('index')
